+++ distance_map.h
@@ -12,12 +12,12 @@ PHP_METHOD(distance_map, score_triangle);
 typedef struct distance_map_object {
     zend_object std;
     coordinate_set_object *coordinate_set;
-    unsigned long **distances;
-    unsigned long size;
+    uint64_t **distances;
+    uint64_t size;
 } distance_map_object;
 
 typedef struct triangle_score {
-    unsigned long x,y,z,row,col;
+    uint64_t x,y,z,row,col;
     struct triangle_score *prev, *next;
 } triangle_score;
 
@@ -30,15 +30,15 @@ void free_distance_map_object(distance_map_object *intern TSRMLS_DC);
 
 void close_gap(distance_map_object *intern, triangle_score *score);
 
-unsigned long score_triangle(distance_map_object *intern, triangle_score *trianlge, int include_gap);
+uint64_t score_triangle(distance_map_object *intern, triangle_score *trianlge, int16_t include_gap);
 
-unsigned long maximum_bound_index_back(distance_map_object *map, unsigned long point, unsigned long *index);
-unsigned long maximum_bound_index_fwrd(distance_map_object *map, unsigned long point, unsigned long *index);
+uint64_t maximum_bound_index_back(distance_map_object *map, uint64_t point, uint64_t *index);
+uint64_t maximum_bound_index_fwrd(distance_map_object *map, uint64_t point, uint64_t *index);
 
-unsigned long skip_up(distance_map_object *map, unsigned long *index, unsigned long required, unsigned long current, int effected_legs);
-unsigned long skip_down(distance_map_object *map, unsigned long *index, unsigned long required, unsigned long current, int effected_legs);
+uint64_t skip_up(distance_map_object *map, uint64_t *index, uint64_t required, uint64_t current, int16_t effected_legs);
+uint64_t skip_down(distance_map_object *map, uint64_t *index, uint64_t required, uint64_t current, int16_t effected_legs);
 
-coordinate_object *get_coordinate(distance_map_object *map, unsigned long index);
+coordinate_object *get_coordinate(distance_map_object *map, uint64_t index);
 int create_distance_map(distance_map_object *map, coordinate_set_object *set);
 
 #ifdef DEBUG_LEVEL
